# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Railgun Business API
  description: |
    Railgun Business Layer API for relay services, enterprise management, and compliance tooling.
    
    **Important:** This API handles ciphertext only. The business layer cannot decrypt message content.
    
    ## Authentication
    - Bearer token (JWT) for user authentication
    - API key for service-to-service
    - OAuth 2.0 for enterprise SSO
    
    ## Rate Limits
    - Free tier: 100 requests/minute
    - Pro tier: 1,000 requests/minute
    - Enterprise: Custom
  version: 1.0.0
  contact:
    name: Railgun API Support
    email: api@railgun.app
  license:
    name: Proprietary
    url: https://railgun.app/terms
servers:
  - url: https://api.railgun.app/v1
    description: Production
  - url: https://api.staging.railgun.app/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Development
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register new account
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Account already exists
  /auth/login:
    post:
      tags: [Auth]
      summary: Login to existing account
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      operationId: refreshToken
      security:
        - refreshToken: []
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout and invalidate tokens
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Logged out successfully

  # ============================================================================
  # Key Management
  # ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
...
    RegisterRequest:
      type: object
      required:
        - identifier
        - identityKey
      properties:
        identifier:
          type: string
          description: Phone number or email
        identityKey:
          type: string
          description: Base64-encoded Ed25519 public key
        deviceName:
          type: string
    LoginRequest:
      type: object
      required:
        - identifier
        - challenge
        - signature
      properties:
        identifier:
          type: string
        challenge:
          type: string
          description: Server-provided challenge
        signature:
          type: string
          description: Challenge signed with identity key
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
        peerId:
          type: string
...